# Refactor rendering and selection to improve performance