--- a/addons/ply/plugin/selector.gd
+++ b/addons/ply/plugin/selector.gd
@@ -52,7 +52,7 @@ func _box_select(camera: Camera3D, v1: Vector2, v2: Vector2, additive: bool) -> 
 		camera.project_position(box2d[3], z_offset),
 	]
 	
-	var planes = []
+	var planes: Array[Plane] = []
 	for i in range(4):
 		var a = box3d[i]
 		var b = box3d[(i+1) % 4]