--- a/addons/ply/resources/ring.gd
+++ b/addons/ply/resources/ring.gd
@@ -1,2 +1,0 @@
-func edge(ply_mesh, edge_idx) -> void:
-	pass