Refactored the PlyEditor node to use an intermediary resource for memoizing frequent and expensive operations. This results in much faster rendering & selection.

For a mesh with 2.3k verts and faces with 4.6k edges seeing improvements on the order of:
rendering: 35ms -> 9ms
ray cast: 10ms -> 5ms (edges continue to be slow but still gains from ~50ms -> 20ms)
box select: 10 SECONDS -> 10ms for verts, 60ms for edges, 150ms for faces

Further performance improvements can be gained by invalidating caches for only mutated geometry, rather than nuking everything on every change.